requirements.txt
google-genai
openai>=1.0.0
pydantic>=2.0.0
arxiv
astroquery
feedparser
beautifulsoup4
requests
pytz